+++ code_examples/core_features/getting_started/02_connect.ts
@@ -3,9 +3,5 @@ import type { ApiPromise } from '@polkadot/api'
 import * as Kilt from '@kiltprotocol/sdk-js'
 
 export async function main(): Promise<ApiPromise> {
-  await Kilt.connect('wss://spiritnet.kilt.io/')
-
-  const api = Kilt.ConfigService.get('api')
-
-  return api
+  return Kilt.connect('wss://spiritnet.kilt.io/')
 }
