+++ code_examples/core_features/getting_started/06_verify_attestation.ts
@@ -18,9 +18,5 @@ export async function main(
   )
 
   // Return false if attestation.revoked is true, or true otherwise.
-  const revokedStatus = !attestationInfo.revoked
-
-  console.log('Revoked status of the attestation', revokedStatus)
-
-  return revokedStatus
+  return !attestationInfo.revoked
 }
