+++ code_examples/core_features/getting_started/07_verify_credential.ts
@@ -3,10 +3,7 @@ import * as Kilt from '@kiltprotocol/sdk-js'
 export async function main(credential: Kilt.ICredential): Promise<void> {
   try {
     await Kilt.Credential.verifyCredential(credential)
-    console.log(
-      "John Doe's credential is valid!",
-      JSON.stringify(credential, null, 2)
-    )
+    console.log("John Doe's credential is valid!")
   } catch {
     console.log("John Doe's credential is not valid.")
   }
