+++ code_examples/core_features/getting_started/index.ts
@@ -6,8 +6,8 @@ import { main as fetchEndpointData } from './05_fetch_endpoint_data'
 import { main as fetchJohnDoeDid } from './03_fetch_did'
 import { main as fetchJohnDoeEndpoints } from './04_fetch_endpoints'
 import { main as printHelloWorld } from './01_print_hello_world'
-import { main as verifyAttestation } from './07_verify_attestation'
-import { main as verifyCredential } from './06_verify_credential'
+import { main as verifyAttestation } from './06_verify_attestation'
+import { main as verifyCredential } from './07_verify_credential'
 
 export async function runAll(): Promise<void> {
   await printHelloWorld()
