+++ code_examples/dapp/src/dapp/domainLinkageCtype.ts
