+++ code_examples/dapp/src/dapp/formatCredential.ts
