+++ code_examples/dapp/src/dapp/sessionCheck.ts
