+++ code_examples/dapp/src/verifier/createRequestCredentialMessage.ts
