+++ code_examples/dapp/src/verifier/decryptCredentialMessage.ts
