+++ code_examples/dapp/src/verifier/encryptRequestCredentialMessage.ts
