+++ code_examples/dapp/src/verifier/generateChallenge.ts
