+++ code_examples/dapp/src/verifier/verifyCredential.ts
