+++ code_examples/dapp/tsconfig.json
@@ -3,8 +3,5 @@
     "moduleResolution": "node",
     "skipLibCheck": true,
     "target": "ES2020"
-  },
-  "exclude": [
-    "src/test.ts"
-  ]
+  }
 }
