+++ code_examples/workshop/.prettierrc
@@ -1,0 +1,6 @@
+{
+  "semi": false,
+  "trailingComma": "none",
+  "singleQuote": true,
+  "printWidth": 80
+}
