+++ code_examples/workshop/attester/generateAccount.ts
