+++ code_examples/workshop/attester/generateCtype.ts
