+++ code_examples/workshop/attester/generateDid.ts
