+++ code_examples/workshop/claimer/generateCredential.ts
