+++ code_examples/workshop/claimer/generateKeypairs.ts
