+++ code_examples/workshop/claimer/generateLightDid.ts
