+++ code_examples/workshop/package.json
@@ -1,0 +1,32 @@
+{
+  "author": "KILT Protocol",
+  "name": "kilt-workshop",
+  "license": "MIT",
+  "scripts": {
+    "check-ts": "yarn tsc -p tsconfig.json --noEmit",
+    "lint": "eslint . --ext .ts --format=codeframe",
+    "lint:fix": "yarn lint --fix",
+    "style": "prettier --check --config .prettierrc '**/*.ts'",
+    "style:fix": "yarn style --write",
+    "fix": "yarn lint:fix && yarn style:fix",
+    "test": "ts-node test.ts"
+  },
+  "dependencies": {
+    "@kiltprotocol/sdk-js": "0.30.0",
+    "dotenv": "^16.0.0"
+  },
+  "devDependencies": {
+    "@types/node": "^17.0.21",
+    "@typescript-eslint/eslint-plugin": "^5.36.0",
+    "@typescript-eslint/parser": "^5.36.0",
+    "eslint": "^8.10.0",
+    "eslint-config-prettier": "^8.4.0",
+    "eslint-formatter-codeframe": "^7.32.1",
+    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "prettier": "^2.5.1",
+    "ts-node": "^10.6.0",
+    "typescript": "^4.7.0"
+  }
+}
