+++ code_examples/workshop/test.ts
@@ -1,3 +1,7 @@
+import { config as envConfig } from 'dotenv'
+import { setTimeout } from 'timers/promises'
+
+import { BN, hexToU8a } from '@polkadot/util'
 import { mnemonicGenerate } from '@polkadot/util-crypto'
 
 import * as Kilt from '@kiltprotocol/sdk-js'
@@ -10,17 +14,15 @@ import { generateKeypairs as generateAttesterKeypairs } from './attester/generat
 import { generateKeypairs as generateClaimerKeypairs } from './claimer/generateKeypairs'
 import { generateCredential } from './claimer/generateCredential'
 import { generateLightDid } from './claimer/generateLightDid'
-import { getFunds } from '../getFunds'
 import { verificationFlow } from './verify'
 
-export async function testWorkshop(
-  faucetAccount: Kilt.KeyringPair,
-  wssAddress: string
-) {
-  console.log('Running the workshop!')
+const SEED_ENV = 'FAUCET_SEED'
 
+async function testWorkshop() {
+  envConfig()
+  process.env.WSS_ADDRESS = 'wss://peregrine.kilt.io/parachain-public-ws'
   Kilt.ConfigService.set({ submitTxResolveOn: Kilt.Blockchain.IS_IN_BLOCK })
-  const api = await Kilt.connect(wssAddress)
+  const api = await Kilt.connect(process.env.WSS_ADDRESS)
 
   // Setup attester account.
   const { account: attesterAccount } = await generateAccount()
@@ -35,7 +37,39 @@ export async function testWorkshop(
     name: 'Karl'
   })
 
-  await getFunds(api, faucetAccount, attesterAccount.address, 5)
+  const faucetSeed = process.env[SEED_ENV]
+  if (!faucetSeed) {
+    console.log(
+      `Account seed with sufficient balance is required. Set the secret seed using the ${SEED_ENV} environment variable.`
+    )
+    throw 'Account seed is missing'
+  }
+
+  const faucetAccount = Kilt.Utils.Crypto.makeKeypairFromSeed(
+    hexToU8a(faucetSeed),
+    'sr25519'
+  )
+
+  const tx = api.tx.balances.transfer(
+    attesterAccount.address,
+    Kilt.BalanceUtils.convertToTxUnit(new BN(5), 0)
+  )
+  try {
+    await Kilt.Blockchain.signAndSubmitTx(tx, faucetAccount)
+  } catch {
+    // Try a second time after a small delay and fetching the right nonce.
+    const waitingTime = 12_000 // 12 seconds
+    console.log(
+      `First submission failed for workshop. Waiting ${waitingTime} ms before retrying.`
+    )
+    await setTimeout(waitingTime)
+    console.log('Retrying...')
+    // nonce: -1 tells the client to fetch the latest nonce by also checking the tx pool.
+    const resignedBatchTx = await tx.signAsync(faucetAccount, { nonce: -1 })
+    await Kilt.Blockchain.submitSignedTx(resignedBatchTx)
+  }
+
+  console.log('Successfully transferred tokens')
 
   // Create attester DID & ensure CType.
   const { fullDid: attesterDid, mnemonic: attesterMnemonic } =
@@ -67,3 +101,13 @@ export async function testWorkshop(
     keyUri: `${lightDid.uri}${lightDid.authentication[0].id}`
   }))
 }
+
+;(async () => {
+  try {
+    await testWorkshop()
+    process.exit(0)
+  } catch (e) {
+    console.error(e)
+    process.exit(1)
+  }
+})()
