+++ code_examples/workshop/verify.ts
