+++ package.json
@@ -5,29 +5,33 @@
   "scripts": {
     "docusaurus": "docusaurus",
     "start": "docusaurus start",
-    "build": "yarn check-ts && yarn generate-examples && yarn check-links && docusaurus build",
+    "build": "yarn check-ts && yarn generate-examples && yarn package-code && yarn check-links && docusaurus build",
     "check-links": "yarn markdown-link-check -c markdown-link-check.config.json -q $(find docs -name '*.md')",
-    "check-ts": "yarn check-ts:core_features && yarn check-ts:vitejs && yarn check-ts:examples",
+    "check-ts": "yarn check-ts:core_features && yarn check-ts:dapp_examples && yarn check-ts:vitejs && yarn check-ts:workshop",
     "check-ts:core_features": "cd code_examples/core_features && yarn install && yarn check-ts",
+    "check-ts:dapp_examples": "cd code_examples/dapp && yarn install && yarn check-ts",
     "check-ts:vitejs": "cd code_examples/vitejs && yarn install && yarn check-ts",
-    "check-ts:examples": "cd code_examples/sdk_examples && yarn install && yarn check-ts",
-    "install": "yarn install:core_features && yarn install:examples && yarn install:scripts && yarn install:vitejs && yarn install:staking && yarn install:examples",
+    "check-ts:workshop": "cd code_examples/workshop && yarn install && yarn check-ts",
+    "install": "yarn install:core_features && yarn install:dapp && yarn install:scripts && yarn install:vitejs && yarn install:staking && yarn install:workshop",
     "install:core_features": "cd code_examples/core_features && yarn install",
+    "install:dapp": "cd code_examples/dapp && yarn install",
     "install:vitejs": "cd code_examples/vitejs && yarn install",
     "install:scripts": "cd scripts && yarn install",
     "install:staking": "cd code_examples/staking && yarn install",
-    "install:examples": "cd code_examples/sdk_examples && yarn install",
-    "fix": "yarn fix:core_features && yarn fix:scripts && yarn fix:staking && yarn fix:vitejs && yarn fix:examples",
+    "install:workshop": "cd code_examples/workshop && yarn install",
+    "fix": "yarn fix:core_features && yarn fix:dapp_examples && yarn fix:scripts && yarn fix:staking && yarn fix:vitejs && yarn fix:workshop",
     "fix:core_features": "cd code_examples/core_features && yarn fix",
+    "fix:dapp_examples": "cd code_examples/dapp && yarn fix",
     "fix:vitejs": "cd code_examples/vitejs && yarn fix",
-    "fix:scripts": "cd code_examples/sdk_examples && yarn fix",
+    "fix:scripts": "cd code_examples/workshop && yarn fix",
     "fix:staking": "cd code_examples/staking && yarn fix",
-    "fix:examples": "cd code_examples/sdk_examples && yarn fix",
+    "fix:workshop": "cd code_examples/workshop && yarn fix",
     "swizzle": "docusaurus swizzle",
-    "deploy": "docusaurus deploy",
+    "deploy": "yarn run package-code && docusaurus deploy",
     "clear": "docusaurus clear",
     "generate-examples": "cd scripts && yarn install && yarn generate",
     "serve": "docusaurus serve",
+    "package-code": "node ./src/utilities/archive.js --folders workshop",
     "write-translations": "docusaurus write-translations",
     "write-heading-ids": "docusaurus write-heading-ids"
   },
